explain "internal judgments" in the topos logic
It's impredicative, but this is a mere formalization of what goes on in
the meaning explanations all the way back to at least Heyting. For
instance, to know "P prop" is to furnish another judgment "P verified"
and its meaning, which is in general an impredicative construction.

I think that what Mark van Atten has said about impredicativity is
relevant, which is that predicativity is not an intrinsic constraint on
constructions by the creating subject, but is rather something that may
be observed *about* constructions by the creating subject. MvA was
talking about the issue of "restrictions on restrictions" on choice
sequences and the apparent impredicativity, but I think it is relevant
here as well.